Extract aggregation code to sub-package
