+++ app/src/main/java/com/samco/trackandgraph/functionslib/DurationAggregationFunction.kt
@@ -17,13 +17,29 @@
 
 package com.samco.trackandgraph.functionslib
 
-import com.samco.trackandgraph.database.entity.AggregatedDataPoint
-import com.samco.trackandgraph.database.entity.DataPointInterface
-import kotlinx.coroutines.yield
+import com.samco.trackandgraph.functionslib.aggregation.FixedBinAggregator
 import org.threeten.bp.Duration
 import org.threeten.bp.OffsetDateTime
 import org.threeten.bp.temporal.TemporalAmount
 
+/**
+ * Add up all data points per plotTotalTime. For example if the plot total time is 1 day and the
+ * sample data contains 3 data points {1, 3, 7} all tracked on the same day then the function will
+ * return a data sample containing 1 point with the value 11.
+ *
+ * The currently supported plotTotalTime values are: Duration.ofHours(1), Period.ofDays(1),
+ * Period.ofWeeks(1), Period.ofMonths(1), Period.ofYears(1)
+ *
+ * If sampleDuration is provided then totals will be generated at least as far back as now minus the
+ * sampleDuration. However if there is more data before this period then it that data will also be
+ * totalled. For clipping see clipDataSample.
+ *
+ * If end time is provided then totals will be generated at least up to this time. However if there
+ * is more data after the end time in the input sample then that data will also be totalled. For
+ * clipping see clipDataSample.
+ *
+ * sampleData.dataPoints should be sorted from oldest timestamp to newest timestamp
+ */
 class DurationAggregationFunction(
     private val timeHelper: TimeHelper,
     private val featureId: Long,
@@ -31,88 +47,9 @@ class DurationAggregationFunction(
     private val endTime: OffsetDateTime?,
     private val binSize: TemporalAmount,
 ) : DataSampleFunction {
-
-    /**
-     * Aggregate all data points into binSize length bins. For example if the bin size is 1 day and the
-     * sample data contains 3 data points {1, 3, 7} all tracked on the same day then the function will
-     * return a data sample with all three data-points as its parent.
-     *
-     * The currently supported binSize values are: Duration.ofHours(1), Period.ofDays(1),
-     * Period.ofWeeks(1), Period.ofMonths(1), Period.ofYears(1)
-     *
-     * If sampleDuration is provided then totals will be generated at least as far back as now minus the
-     * sampleDuration. However if there is more data before this period then it that data will also be
-     * totalled. For clipping see clipDataSample.
-     *
-     * If end time is provided then totals will be generated at least up to this time. However if there
-     * is more data after the end time in the input sample then that data will also be totalled. For
-     * clipping see clipDataSample.
-     *
-     * sampleData.dataPoints should be sorted from oldest timestamp to newest timestamp
-     */
     override suspend fun execute(dataSample: DataSample): DataSample {
-        val newData = mutableListOf<AggregatedDataPoint>()
-
-        val firstDataPointTime =
-            dataSample.dataPoints.firstOrNull()?.cutoffTimestampForAggregation()
-        val lastDataPointTime = dataSample.dataPoints.lastOrNull()?.cutoffTimestampForAggregation()
-
-        val latest = getEndTimeNowOrLatest(lastDataPointTime, endTime)
-        val earliest = getStartTimeOrFirst(firstDataPointTime, latest, endTime, sampleDuration)
-        var currentTimeStamp = timeHelper.findBeginningOfTemporal(earliest, binSize).minusNanos(1)
-        var index = 0
-        while (currentTimeStamp.isBefore(latest)) {
-            currentTimeStamp = currentTimeStamp.with { ld -> ld.plus(binSize) }
-            val points = dataSample.dataPoints.drop(index)
-                .takeWhile { dp ->
-                    dp.cutoffTimestampForAggregation()
-                        .isBefore(currentTimeStamp)
-                }
-            index += points.size
-            newData.add(
-                AggregatedDataPoint(
-                    timestamp = currentTimeStamp,
-                    value = Double.NaN, // this value gets overwritten when calling a function on RawAggregatedDatapoints
-                    featureId = featureId,
-                    parents = points
-                )
-            )
-            yield()
-        }
-        return RawAggregatedDatapoints(newData).sum()
-    }
-
-    private fun getStartTimeOrFirst(
-        firstDataPointTime: OffsetDateTime?,
-        latest: OffsetDateTime,
-        endTime: OffsetDateTime?,
-        sampleDuration: Duration?
-    ): OffsetDateTime {
-        val beginningOfDuration = sampleDuration?.let { endTime?.minus(it) }
-        val durationBeforeLatest = sampleDuration?.let { latest.minus(it) }
-        return listOf(
-            firstDataPointTime,
-            beginningOfDuration,
-            durationBeforeLatest,
-            latest
-        ).minBy { t -> t ?: OffsetDateTime.MAX }!!
-    }
-
-    private fun getEndTimeNowOrLatest(
-        lastDataPointTime: OffsetDateTime?,
-        endTime: OffsetDateTime?
-    ): OffsetDateTime {
-        val now = OffsetDateTime.now()
-        return when {
-            //last == null && endTime == null -> now
-            endTime == null -> listOf(lastDataPointTime, now).maxBy { t ->
-                t ?: OffsetDateTime.MIN
-            }!!
-            else -> listOf(lastDataPointTime, endTime).maxBy { t -> t ?: OffsetDateTime.MIN }!!
-        }
-    }
-
-    private fun DataPointInterface.cutoffTimestampForAggregation(): OffsetDateTime {
-        return timestamp - timeHelper.aggregationPreferences.startTimeOfDay
+        return FixedBinAggregator(timeHelper, featureId, sampleDuration, endTime, binSize)
+            .aggregate(dataSample)
+            .sum()
     }
 }
